Move the ModuleValidator from /wtf to /app
